requirements.txt
django
django-admin-interface
djangorestframework-simplejwt
djangorestframework
markdown
django-filter
requests
gunicorn
supervisor
django-environ
sendgrid
django-simple-captcha
django-cors-headers
Faker
PyMySQL
mysqlclient
django-redis
django-redis-sessions
django-restql
drf-yasg
rich
ShopifyAPI